requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
mangum==0.17.0
kecam==1.3.17
keras_cv_attention_models==1.3.9
lion_pytorch==0.1.2
pytorch_lightning==2.2.0.post0
torchsummary==1.5.1